// src/core/store/type/book/types.ts
enum EBookAction {
	GET_BOOKS = 'GET_BOOKS',
	SAGA_GET_BOOKS = 'SAGA_GET_BOOKS',

	GET_BOOKS_CATEGORIES = 'GET_BOOKS_CATEGORIES',
	SAGA_GET_BOOKS_CATEGORIES = 'SAGA_GET_BOOKS_CATEGORIES',

	ADD_BOOK_TO_CART = 'ADD_BOOK_TO_CART',
	REMOVE_BOOK_FROM_CART = 'REMOVE_BOOK_FROM_CART',
	DESTROY_BOOK_FROM_CART = 'DESTROY_BOOK_FROM_CART',

	CHANGE_BALANCE = 'CHANGE_BALANCE',
	CLEAR_CART = 'CLEAR_CART',

	CHANGE_FILTER = 'CHANGE_FILTER',
	SAGA_GET_BOOKS_FILTER = 'SAGA_GET_BOOKS_FILTER',
}

export default EBookAction